# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Planes",
    "version": "1.0.0"
  },
  "components": {
    "schemas": {
...
      "PlanePurchaseRequest": {
        "type": "object",
        "properties": {
          "tailNumber": {
            "type": "string",
            "pattern": "^[0-9A-Z]{6}$"
          },
          "model": {
            "type": "string"
          },
          "airport": {
            "type": "string"
          }
        },
        "required": [
          "tailNumber",
          "model",
          "airport"
        ],
        "additionalProperties": false
      },
      "PlaneRefuelRequest": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "format": "float",
            "minimum": 0
          }
        },
        "required": [
          "amount"
        ],
        "additionalProperties": false
      },
      "PlaneEmbarkRequest": {
        "type": "object",
        "properties": {
          "destination": {
            "type": "string"
          }
        },
        "required": [
          "destination"
        ],
        "additionalProperties": false
      },
      "PlaneResponse": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "owner": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "whereabouts": {
            "anyOf": [
              {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "airport"
                    ]
                  },
                  "airport": {
                    "type": "string"
                  }
                },
                "required": [
                  "type",
                  "airport"
                ],
                "additionalProperties": false,
                "x-required": true
              },
              {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "enRoute"
                    ]
                  },
                  "enRoute": {
                    "type": "object",
                    "properties": {
                      "departure": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "origin": {
                        "type": "string"
                      },
                      "destination": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "departure",
                      "origin",
                      "destination"
                    ],
                    "additionalProperties": false
                  }
                },
                "required": [
                  "type",
                  "enRoute"
                ],
                "additionalProperties": false,
                "x-required": true
              }
            ]
          },
          "fuel": {
            "type": "number",
            "format": "float",
            "minimum": 0
          },
          "upgradeLevel": {
            "type": "integer",
            "minimum": 0,
            "maximum": 3
          },
          "condition": {
            "type": "number",
            "format": "float",
            "minimum": 0,
            "maximum": 100
          }
        },
        "required": [
          "_id",
          "owner",
          "model",
          "whereabouts",
          "fuel",
          "upgradeLevel",
          "condition"
        ],
        "additionalProperties": false
      },
...
    }
  },
  "paths": {
...
    "/planes": {
      "get": {
        "summary": "Get all planes",
        "description": "Returns all planes in the system regardless of owner. Public endpoint for viewing the complete fleet.",
        "tags": [
          "Planes"
        ],
        "responses": {
          "200": {
            "description": "List of all planes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlaneResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Purchase a new plane",
        "description": "Buy a new plane with a user-supplied tail number and place it at the specified airport. Authenticated user becomes the owner. Deducts cost from user's funds. Client should check tail number availability via GET /planes/:id first.",
        "tags": [
          "Planes"
        ],
        "security": [
          {
            "oauth2": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanePurchaseRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Plane purchased successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaneResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request (invalid model, invalid airport, invalid tail number format)"
          },
          "401": {
            "description": "Not authenticated"
          },
          "402": {
            "description": "Insufficient funds"
          },
          "404": {
            "description": "Airport not found"
          },
          "409": {
            "description": "Tail number already exists"
          }
        }
      }
    },
    "/planes/{id}": {
      "get": {
        "summary": "Get a specific plane by tail number",
        "description": "Returns detailed information about a single plane. Public endpoint.",
        "tags": [
          "Planes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Aircraft tail number (e.g., N058DB)",
            "example": "N058DB"
          }
        ],
        "responses": {
          "200": {
            "description": "Plane details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaneResponse"
                }
              }
            }
          },
          "404": {
            "description": "Plane not found"
          }
        }
      },
      "delete": {
        "summary": "Decommission plane",
        "description": "Permanently removes the plane from service with no refund. Requires authentication and plane ownership.",
        "tags": [
          "Planes"
        ],
        "security": [
          {
            "oauth2": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Aircraft tail number",
            "example": "N058DB"
          }
        ],
        "responses": {
          "204": {
            "description": "Plane decommissioned successfully"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not the plane owner"
          },
          "404": {
            "description": "Plane not found"
          }
        }
      }
    },
    "/planes/{id}/upgrade": {
      "put": {
        "summary": "Upgrade plane to next level",
        "description": "Increments upgrade level by 1 (max level 3). Requires authentication and plane ownership. Deducts upgrade cost from user's funds.",
        "tags": [
          "Planes"
        ],
        "security": [
          {
            "oauth2": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Aircraft tail number",
            "example": "N058DB"
          }
        ],
        "responses": {
          "200": {
            "description": "Plane upgraded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaneResponse"
                }
              }
            }
          },
          "400": {
            "description": "Cannot upgrade (already at max level 3)"
          },
          "401": {
            "description": "Not authenticated"
          },
          "402": {
            "description": "Insufficient funds"
          },
          "403": {
            "description": "Not the plane owner"
          },
          "404": {
            "description": "Plane not found"
          }
        }
      }
    },
    "/planes/{id}/embark": {
      "put": {
        "summary": "Depart for destination airport",
        "description": "Start a flight to the specified airport. Requires authentication, plane ownership, and sufficient fuel. Plane must be at an airport (not already en route).",
        "tags": [
          "Planes"
        ],
        "security": [
          {
            "oauth2": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Aircraft tail number",
            "example": "N058DB"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaneEmbarkRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Flight started successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaneResponse"
                }
              }
            }
          },
          "400": {
            "description": "Cannot embark (insufficient fuel, already en route, destination same as current location)"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not the plane owner"
          },
          "404": {
            "description": "Plane or destination airport not found"
          }
        }
      }
    },
    "/planes/{id}/refuel": {
      "patch": {
        "summary": "Add fuel to plane",
        "description": "Refuel the plane by specified amount in kilograms. Requires authentication and plane ownership. Cannot exceed fuel capacity. Deducts fuel cost from user's funds.",
        "tags": [
          "Planes"
        ],
        "security": [
          {
            "oauth2": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Aircraft tail number",
            "example": "N058DB"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaneRefuelRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Plane refueled successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaneResponse"
                }
              }
            }
          },
          "400": {
            "description": "Cannot refuel (would exceed fuel capacity, plane en route)"
          },
          "401": {
            "description": "Not authenticated"
          },
          "402": {
            "description": "Insufficient funds"
          },
          "403": {
            "description": "Not the plane owner"
          },
          "404": {
            "description": "Plane not found"
          }
        }
      }
    },
    "/planes/{id}/repair": {
      "patch": {
        "summary": "Repair plane to 100% condition",
        "description": "Fully repairs the plane to 100% condition. Requires authentication and plane ownership. Deducts repair cost from user's funds.",
        "tags": [
          "Planes"
        ],
        "security": [
          {
            "oauth2": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Aircraft tail number",
            "example": "N058DB"
          }
        ],
        "responses": {
          "200": {
            "description": "Plane repaired successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaneResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "402": {
            "description": "Insufficient funds"
          },
          "403": {
            "description": "Not the plane owner"
          },
          "404": {
            "description": "Plane not found"
          }
        }
      }
    },
...
  },
...
}